EChapterID, ENPCID enum들 별도 파일로 분리
